refactor command to support multiple subcommands
