+++ tool/tctl/common/discovery/nodes_azure.go
@@ -27,8 +27,6 @@ import (
 	libevents "github.com/gravitational/teleport/lib/events"
 )
 
-const cloudAzure = "Azure"
-
 // azureInfo holds Azure-specific instance metadata.
 type azureInfo struct {
 	VMID           string `json:"vm_id"`
