+++ tool/tctl/common/discovery/nodes_render.go
