+++ tool/tctl/common/discovery/nodes_render_test.go
